+++ extensions/builder/artifact/artifact-definition.ts
@@ -1,0 +1,18 @@
+export type ArtifactDefinition = {
+  /**
+   *
+   */
+  name: string;
+
+  /**
+   *
+   */
+  description?: string;
+
+  /**
+   *
+   */
+  globPatterns: string[];
+
+  storageResolver?: string[];
+};
