+++ extensions/builder/artifact/index.ts
@@ -1,0 +1,4 @@
+export { ArtifactDefinition } from './artifact-definition';
+export { ExtensionArtifact } from './extension-artifact';
+export { ArtifactRef } from './ref';
+export { Artifact } from './artifact';
