+++ extensions/builder/artifact/ref.ts
@@ -1,0 +1,1 @@
+export class ArtifactRef {}
