+++ extensions/builder/storage/index.ts
@@ -1,0 +1,1 @@
+export { StorageResolver } from './storage-resolver';
