+++ extensions/builder/storage/storage-resolver.ts
@@ -1,0 +1,5 @@
+import { Artifact } from '../artifact';
+
+export interface StorageResolver {
+  store(artifacts: Artifact[]): Promise<void>;
+}
