+++ extensions/preview/preview.task.ts
@@ -1,6 +1,6 @@
 import { resolve, join } from 'path';
 import { ExecutionContext } from '@teambit/environments';
-import { BuildContext, BuildResults, BuildTask } from '@teambit/builder';
+import { BuildContext, BuildResults, BuildTask, TaskLocation } from '@teambit/builder';
 import { Bundler, BundlerContext, BundlerMain, Target } from '@teambit/bundler';
 import { Compiler } from '@teambit/compiler';
 import { ComponentMap } from '@teambit/component';
@@ -24,6 +24,8 @@ export class PreviewTask implements BuildTask {
   ) {}
 
   id = 'teambit.bit/preview';
+
+  location: TaskLocation = 'end';
 
   async execute(context: BuildContext): Promise<BuildResults> {
     const defs = this.preview.getDefs();
